cleanup lots of warnings and disabled warnings (#250)

+++ external/gfx/imgui/imgui_spinner.cpp
@@ -68,7 +68,7 @@ namespace ImGui {
         window->DrawList->PathClear();
         
         auto num_segments = 30.f;
-        auto start = fabs(float(ImSin(float(g.Time*1.8f))*(num_segments-5)));
+        auto start = (float)fabs(float(ImSin(float(g.Time*1.8f))*(num_segments-5)));
         
         const float a_min = IM_PI*2.0f * start / num_segments;
         const float a_max = IM_PI*2.0f * (num_segments-3) / num_segments;
